Cv feature

+++ backend/src/modules/user/user.service.js
@@ -3,6 +3,7 @@ import Token from "./token.model.js"
 import User from "./user.model.js"
 import { abortTransaction, commitTransaction, startTransaction } from "../../utils/session.js"
 import Upload from "../../shared/upload/upload.model.js"
+import { deleteFileFromCloudinary } from "../../utils/deleteFileFromCloudinary.js"
 
 
 export const getUser = async (email) => {
@@ -104,7 +105,14 @@ export const deleteUserSevice = async (userId) => {
     try {
         const UserDeleted = await User.findOneAndDelete({ _id: userId }, { session })
         const tokensDeleted = await Token.deleteMany({ userId }, { session })
-        const uploadDeleted = await Upload.deleteMany({ userId }, { session })
+        const uploadDeleted = await Upload.find({ userId })
+
+        if(uploadDeleted) {
+            for(const upload of uploadDeleted) {
+                await deleteFileFromCloudinary(upload.publicId);
+            }
+            await Upload.deleteMany({ userId }, { session })
+        }
 
         if(UserDeleted && tokensDeleted && uploadDeleted) {
             await commitTransaction(session)
